issues with colliders still, will try using tag name instead of layer
