+++ Assets/Prefabs/Magician.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: dc60115e30e5dfe4e86f2b9b2455e821
+guid: f21e4d03a769bac479e12c6b6b44abeb
 PrefabImporter:
   externalObjects: {}
   userData:
