--- Assets/Prefabs/PlayerSpriteSheet_8.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 26947c25444665142bc76be054834f6d
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 21300000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
