+++ Assets/Prefabs/PlayerSpriteSide.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 3b85cbd213636dc428edab25f5b21067
-DefaultImporter:
+guid: bf2d5c668ead6764386a03672b2be29e
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
