+++ Assets/Scenes/KongScene.unity.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 552fa98a67be0504f8e086c2a8a86c26
-PrefabImporter:
+guid: f16b0a1464e60e444bb797505275cd97
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
