+++ Assets/Scripts/KongPlayer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4ae8f4d84ac92dc4ea327d64a84dff12
+guid: a62e8174717a97045a32530e095977b9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
